Mtab isn't needed
`e4crypt add_key -S $salt $path` is a misleading command. It will
create a key with the salt, and a separate key for the $path using
the filesystem's s_encrypt_pw_salt field as the salt.

It only uses mtab to generate keys when the user hasn't passed a
salt explicitly, such as the command `e4crypt add_key $path` or
`e4crypt add_key`

So we don't actually need mtab at all